posterscan-backend/requirements.txt
flask
qreader
openai
python-dotenv
beautifulsoup4
requests
gunicorn
easyocr
ics
flask-cors
